Fix formatting, outline, TOC of README
The last submission was a good start.  There were a number of linter violations
that I took care of here.

The overall flow did not feel good either.  Therefore, I changed up all the
headings and location of the information.

Nothing was removed.  However, a number of items were re-written.